Feature/cart

+++ SERVER/routes/cartController.js
@@ -1,20 +1,28 @@
 const { Router } = require("express");
 const Cart = require("../models/cart");
+const CartProduct = require("../models/cartproduct");
 const checkAuth = require("../middlewares/checkAuth");
 const router = new Router();
-
-router.get("/", checkAuth, async (req, res, next) => {
-  const carts = await Cart.findAll({
-    where: req.query,
-  });
-  res.json(carts);
-});
+const { v4: isUUID } = require('uuid');
 
 router.post("/", async (req, res, next) => {
   try {
-    const cart = await Cart.create(req.body);
+    const { id_user } = req.body;
+
+    if (!id_user) {
+      console.error("User ID is required");
+      return res.status(400).json({ error: "User ID is required" });
+    }
+
+    const expire_date = new Date(Date.now() + 15 * 60 * 1000);
+    const cart = await Cart.create({
+      id_user,
+      expire_date,
+    });
+
     res.status(201).json(cart);
   } catch (e) {
+    console.error("Error creating cart:", e);
     next(e);
   }
 });
@@ -22,6 +30,30 @@ router.post("/", async (req, res, next) => {
 router.get("/:id", async (req, res, next) => {
   try {
     const cart = await Cart.findByPk(parseInt(req.params.id));
+    if (cart) {
+      res.json(cart);
+    } else {
+      res.sendStatus(404);
+    }
+  } catch (e) {
+    next(e);
+  }
+});
+
+router.post("/getByIDUser", async (req, res, next) => {
+  try {
+    const id_user = req.body.id_user;
+
+    // Vérifier si l'ID utilisateur est un UUID valide
+    if (!isUUID(id_user)) {
+      return res.status(400).send('Invalid user ID format');
+    }
+
+    const cart = await Cart.findOne({
+       where: {
+         id_user: id_user 
+        } 
+      });
     if (cart) {
       res.json(cart);
     } else {
@@ -68,17 +100,28 @@ router.delete("/:id", async (req, res, next) => {
   }
 });
 
-router.put("/:id", async (req, res, next) => {
+router.put("/", async (req, res, next) => {
   try {
+    const { id_user } = req.body;
+    console.log(`Received request to update cart for user ID: ${id_user}`);
+
+    if (!id_user) {
+      console.error("User ID is required");
+      return res.status(400).json({ error: "User ID is required" });
+    }
+
+    // Supprimez les paniers existants pour cet utilisateur
     const nbDeleted = await Cart.destroy({
       where: {
-        id: parseInt(req.params.id),
+        id_user: id_user,
       },
     });
+
     const cart = await Cart.create({
-      ...req.body,
-      id: parseInt(req.params.id),
+      id_user,
     });
+
+    // Répondez avec le nouveau panier
     res.status(nbDeleted ? 200 : 201).json(cart);
   } catch (e) {
     next(e);
